added some new methods to student-related classes

+++ src/main/java/ua/lviv/iot/spring/rest/controller/StudentsController.java
@@ -1,59 +1,61 @@
 package ua.lviv.iot.spring.rest.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ua.lviv.iot.spring.rest.business.StudentService;
 import ua.lviv.iot.spring.rest.model.Student;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
 
 @RequestMapping("/students")
 @RestController
 public class StudentsController {
 
-    private Map<Integer, Student> students = new HashMap<>();
-    private AtomicInteger IdCounter = new AtomicInteger();
 
     @Autowired
     private StudentService studentService;
 
 
     @GetMapping
-    public List<Student> getAllStudents() {
-        return studentService.findAll();
+    public List<Student> getStudents(final @RequestParam(value = "firstName", required = false) String firstName) {
+        if (firstName == null) {
+            return studentService.getAllStudents();
+        }
+        return studentService.getAllStudentsByFirstName(firstName);
     }
 
     @GetMapping(path = "/{id}")
     public Student getStudent(final @PathVariable("id") Integer studentId) {
-        return students.get(studentId);
+        return studentService.getStudentById(studentId);
     }
 
     @PostMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
     public Student createStudent(final @RequestBody Student student) {
-        student.setId(IdCounter.incrementAndGet());
-        students.put(student.getId(), student);
-        studentService.createStudent(student);
-        return student;
+        return studentService.createStudent(student);
     }
 
     @DeleteMapping(path = "/{id}")
     public ResponseEntity<Student> deleteStudent(final @PathVariable("id") Integer studentId) {
-        HttpStatus status = students.remove(studentId) == null ? HttpStatus.NOT_FOUND : HttpStatus.OK;
-        studentService.deleteStudent(studentId);
-        return ResponseEntity.status(status).build();
+        Student studentToCheck = studentService.getStudentById(studentId);
+        if (studentToCheck != null) {
+            studentService.deleteStudent(studentId);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PutMapping(path = "/{id}")
     public ResponseEntity<Student> updateStudent(final @PathVariable("id") Integer studentId, final @RequestBody Student student) {
-        HttpStatus status = students.put(studentId, student) == null ? HttpStatus.NOT_FOUND : HttpStatus.OK;
-        studentService.updateStudent(studentId, student);
-        return ResponseEntity.status(status).build();
+        Student studentToCheck = studentService.getStudentById(studentId);
+        if (studentToCheck != null) {
+            studentService.updateStudent(studentId, student);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
